* WebSocket for both stream and control

+++ src/cgvhid/cgvhid_client/cgvhid.cpp
@@ -188,8 +188,8 @@ int Cgvhid::AbsoluteMouseUpdate(std::uint8_t button,
 }
 
 int Cgvhid::RelativeMouseUpdate(std::uint8_t button,
-                                std::int8_t x,
-                                std::int8_t y,
+                                std::int16_t x,
+                                std::int16_t y,
                                 std::uint8_t hwheel,
                                 std::uint8_t vwheel) noexcept {
   // Enterprise version
